src/LyxBlog/account.py
#! /usr/bin/env python
# -*- coding: utf-8 -*-
#####################       A U T H O R       ##########################
#                                                                      #
#   Copyright 2010 Jack Desert                                         #
#   <jackdesert556@gmail.com>                                          #
#   <http://www.LetsEATalready.com>                                    #
#                                                                      #
######################      L I C E N S E     ##########################
#                                                                      #
#   This file is part of LyXBlogger.                                   #
#                                                                      #
#   LyXBlogger is free software: you can redistribute it and/or modify #
#   it under the terms of the GNU General Public License as published  #
#   by the Free Software Foundation, either version 3 of the License,  #
#   or (at your option) any later version.                             #
#                                                                      #
#   LyXBlogger is distributed in the hope that it will be useful,      #
#   but WITHOUT ANY WARRANTY; without even the implied warranty of     #
#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the      #
#   GNU General Public License for more details.                       #
#                                                                      #
#   You should have received a copy of the GNU General Public License  #
#   along with LyXBlogger.  If not, see <http://www.gnu.org/licenses>. #
#                                                                      #
########################################################################

class Account:
    def __init__(self, section_id, url, username, password):
        self.__url = url
        self.__username = username
        self.__password = password 
        self.__section_id = section_id

    def __eq__(self, other):
        return self.__url == other._Account__url and self.__username == other._Account__username and self.__password == other._Account__password

    def set_password(self, password):
        self.__password = password

    def get_section_id(self):
        return self.__section_id

    def get_password(self):
        return self.__password

    def get_url(self):
        return self.__url
    def get_username(self):
        return self.__username

